[#34] Move some functions out of drawing.ts
